Named commands (read desc)
Added named commands for scoring coral and intaking coral to pathplanner and to robotcontainer.

Still working on an auto-align named command that will run during auto and only use a single-tag tracking. In pathplanner, this will run DURING a path as we're only changing the poses we send to the swerve drive. I'm gonna talk to Liam abt this during practice.